pyproject.toml
[tool.poetry]
name = "mlops-test-repo"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = "^1.4.2"
matplotlib = "^3.5.2"
tensorflow = "^2.8.0"
click = "^8.1.3"
numpy = "^1.22.3"
dvc = {extras = ["all"], version = "^2.10.2"}
scipy = "^1.8.0"
pywin32 = "^304"
mlflow = "^1.25.1"
python-dotenv = "^0.20.0"
uvicorn = "^0.17.6"
fastapi = "^0.78.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
